backend/requirements.txt
# Food Vision Pro Backend Dependencies

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis[hiredis]==5.0.1

# HTTP client
aiohttp==3.9.1
httpx==0.25.2

# Image processing and ML
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
ultralytics==8.0.196
onnxruntime==1.16.3
torch==2.1.1
torchvision==0.16.1

# Data processing
PyYAML==6.0.1
aiofiles==23.2.1

# Optional: S3/MinIO storage
boto3==1.34.0
minio==7.2.0

# Optional: Stripe billing
stripe==7.8.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
ruff==0.1.6
mypy==1.7.1

# Logging and monitoring
structlog==23.2.0
